pyproject.toml
[tool.poetry]
name = "mattermost-data-warehouse"
version = "0.1.0"
description = "A set of tools for managing data at Mattermost's data warehouse"
authors = ["Munish Sharma <munish.sharma@mattermost.com>", "Ioannis Foukarakis <ioannis.foukarakis@mattermost.com>"]
readme = "README.md"
packages = [
    {include = "extract"},
    {include = "utils"},
]
exclude = [
    "dags",
    "dags/billing/*",
    "plugins",
]

[tool.poetry.dependencies]
python = "^3.10"
certifi = "2018.8.24"
pandas = "^2.1.4"
requests = "2.28.1"
user-agents = "2.1"
snowflake-sqlalchemy = "^1.5.1"
snowflake-connector-python = "^3.6.0"

sqlalchemy = "<2.0.0"
ghapi = "^1.0.3"
click = "^8.1.3"
tabulate = "^0.9.0"
pydantic = "^2.8.2"
pyyaml = "^6.0.2"
boto3 = "^1.35.35"
looker-sdk = "^25.0.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.4"
pytest-responses = "^0.5.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
flake8-bugbear = "^22.10.27"
flake8-comprehensions = "^3.10.1"
flake8-simplify = "^0.19.3"
flake8-eradicate = "^1.4.0"
responses = "0.22.0"
mock = "^4.0.3"
pytest = "^7.1.3"
pytest-freezegun = "^0.4.2"
moto = {extras = ["s3"], version = "^5.0.16"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rudder = "utils.rudderstack.__main__:rudder"
contributors = "utils.github.__main__:contributors"
push_proxy = "utils.push_proxy.__main__:push_proxy"
geolite = "utils.geolite.__main__:geolite"
packets = "utils.packets.__main__:packets"
snowflake = "utils.db.__main__:snowflake"


[tool.pytest.ini_options]
addopts = "--ignore=airflow"
testpaths = [ "tests" ]

[tool.coverage.run]
omit = [".*", "*/site-packages/*", "tests/*"]
source = ["extract", "utils", "tests"]
branch = true

[tool.coverage.report]
show_missing = true
precision = 2

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310']
skip-string-normalization = true

[tool.isort]
append_only = true
line_length = 120
combine_as_imports = true
default_section = "THIRDPARTY"
skip = ["build", "k8s", "load", "transform", ".tox", ".venv", ".venv-airflow"]
skip_glob = ["*.pyi"]
profile = "black"
